+++ changelog.md
@@ -6,7 +6,7 @@
   ([#379](https://github.com/feltjs/gro/pull/379))
 - add `gen` plugin that defaults to watch mode
   ([#283](https://github.com/feltjs/gro/pull/283))
-- add `--reset` flag to `gro deploy`
+- add `--reset` and `--origin` flags to `gro deploy`
   ([#383](https://github.com/feltjs/gro/pull/383))
 
 ## 0.79.0